+++ pixelfed/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Version bump for new base image
 * Set .env variable "FILESYSTEM_DRIVER=s3"
+* Switch to latest package stream
+* Build as non-layered image
 
 ---